Re #4 Fix a bunch of model warnings

+++ sansmodels/src/sans/models/c_models/SquareWell.cpp
@@ -65,7 +65,7 @@ double SquareWellStructure :: operator()(double q) {
 	double norm = 0.0;
 
 	// Loop over radius weight points
-	for(int i=0; i<weights_rad.size(); i++) {
+	for(size_t i=0; i<weights_rad.size(); i++) {
 		dp[0] = weights_rad[i].value;
 
 		sum += weights_rad[i].weight
